requirements.txt
# Pinned dependencies for MEF Portal
# Generated for reproducible environments

# Flask and extensions
flask==2.3.3
flask-cors==4.0.0
flask-wtf==1.2.1
flask-login==0.6.3
flask-limiter==3.5.0

# Database
mysql-connector-python==8.2.0

# PDF generation
fpdf2==2.7.5

# Security
werkzeug==2.3.7
bleach==6.1.0
PyJWT==2.8.0
cryptography==41.0.7

# Testing (optional but recommended)
pytest==7.4.3
pytest-flask==1.3.0
